Fixed issue #3124
