+++ participant-datastore/consent-mgmt-module/consent-mgmt/src/main/java/com/google/cloud/healthcare/fdamystudies/MyStudiesUserRegWsApplication.java
@@ -13,7 +13,9 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableScheduling
 @SpringBootApplication
 @EntityScan("com.google.cloud.healthcare.fdamystudies.model")
 @EnableAutoConfiguration(exclude = HibernateJpaAutoConfiguration.class)
