+++ participant-datastore/consent-mgmt-module/consent-mgmt/src/main/java/com/google/cloud/healthcare/fdamystudies/config/AppConfig.java
@@ -8,7 +8,22 @@
 
 package com.google.cloud.healthcare.fdamystudies.config;
 
+import com.google.cloud.storage.Storage;
+import com.google.cloud.storage.StorageOptions;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @Configuration
-public class AppConfig {}
+public class AppConfig extends CommonModuleConfiguration {
+
+  @Bean
+  public BCryptPasswordEncoder passwordEncoder() {
+    return new BCryptPasswordEncoder();
+  }
+
+  @Bean
+  public Storage storageService() {
+    return StorageOptions.getDefaultInstance().getService();
+  }
+}
