--- participant-datastore/consent-mgmt-module/consent-mgmt/src/main/java/com/google/cloud/healthcare/fdamystudies/config/AuthenticationEntryPointImpl.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Use of this source code is governed by an MIT-style
- * license that can be found in the LICENSE file or at
- * https://opensource.org/licenses/MIT.
- */
-
-package com.google.cloud.healthcare.fdamystudies.config;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.www.BasicAuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-@Component
-public class AuthenticationEntryPointImpl extends BasicAuthenticationEntryPoint {
-
-  @Override
-  public void commence(
-      HttpServletRequest request, HttpServletResponse response, AuthenticationException authEx)
-      throws IOException {
-    response.addHeader("WWW-Authenticate", "Basic realm=" + getRealmName());
-    response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-    PrintWriter writer = response.getWriter();
-    writer.println("HTTP Status 401 - " + authEx.getMessage());
-  }
-
-  @Override
-  public void afterPropertiesSet() {
-    setRealmName("participant-consent-datastore");
-    super.afterPropertiesSet();
-  }
-}
