--- participant-datastore/consent-mgmt-module/consent-mgmt/src/main/java/com/google/cloud/healthcare/fdamystudies/config/BeanConfig.java
@@ -1,41 +1,0 @@
-/*
-
-* Copyright 2020 Google LLC
-*
-* Use of this source code is governed by an MIT-style
-* license that can be found in the LICENSE file or at
-* https://opensource.org/licenses/MIT.
-*/
-
-package com.google.cloud.healthcare.fdamystudies.config;
-
-import com.google.cloud.storage.Storage;
-import com.google.cloud.storage.StorageOptions;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableScheduling;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@EnableScheduling
-@Configuration
-@EnableWebMvc
-public class BeanConfig extends CommonModuleConfiguration {
-
-  @Bean
-  public Storage storageService() {
-    return StorageOptions.getDefaultInstance().getService();
-  }
-
-  @Bean
-  public WebMvcConfigurer corsConfigurer() {
-
-    return new WebMvcConfigurer() {
-      @Override
-      public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**").allowedOrigins("*").allowedHeaders("*").allowedMethods("*");
-      }
-    };
-  }
-}
